Implemented alarm functionality!
The alarm now works!
After setting the alarm hour and time, the program checks if the current hour and minute are equal to the alarm and minute of the set alarm, and if so, prints "Alarm!"

Yeah, so no audio playing yet, and I don't have the display flashing or doing anything fancy yet either, but it's a start!