Implementing more GPIO stuff
Added falling event detection for all 7 pins and added corresponding functions.

....I'm going to need 7 more functions for when they get pushed down. Ugh.
There has to be a better way to do this.

Like, listen for a rising or falling voltage on all pins, and just have one functions that returns what pin the activity is on, and if it's rising or falling. That would make things way cleaner. Maybe I can make something like that.