Fixed syntax + tabbing errors
Fixed a few syntax errors and weird tabbing/spacing that wasn't agreeing with python.
Also added a time.sleep() event where one should have been in order to show the alarm time to the user before having them confirm it. And, when it's time for the user to take their pills, it will now say "Monday" instead of "Mon," etc. Woo!


Now to test this on the Raspberry Pi and see if:
A. I (hopefully don't) get any runtime errors from using the RPi.GPIO library in a way that is wrong
B. The hardware actually goddamn works!! It better!